Refactor the alarm feature
