--- testvolume.py
@@ -1,7 +1,0 @@
-from utils import play_music
-
-print('正在播放音樂 (按 ctrl + c 可停止)...')
-
-play_music()
-
-print('播放完畢')
